Modified the classes structure in base at the v3 version documentation (Step-2), added the Section20 management. The section 3 is still incomplete, porting not terminated to the v3 version documentation

+++ PSE.Model/Output/Interfaces/IBase.cs
@@ -12,7 +12,7 @@ namespace PSE.Model.Output.Interfaces
 
         string? Description { get; set; }
 
-        int? ValorNumber { get; set; }
+        long? ValorNumber { get; set; }
 
         string? Isin { get; set; }
